+++ lib/src/model/user.dart
@@ -21,5 +21,6 @@ class User extends JsonObject {
 
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
 
+  @override
   Map<String, dynamic> toJson() => _$UserToJson(this);
 }